Clean the cache, bc it caused some problems

+++ csharp/ASP.NET/HoppKutya/api/hoppKutya/Models/GameTables.cs
@@ -8,8 +8,8 @@ namespace hoppKutya.Models
 		public int TableId { get; set; }
 		public string? Player1 { get; set; }
 		public string? Player2 { get; set; }
-		public int Player1Points { get; set; }
-		public int Player2Points { get; set; }
+		public double Player1Points { get; set; }
+		public double Player2Points { get; set; }
 		public DateTime LastHoppKutya { get; set; }
 	}
 }
